+++ src/consumer/component/package-json-utils.js
@@ -17,6 +17,8 @@ import JSONFile from './sources/json-file';
 import npmRegistryName from '../../utils/bit/npm-registry-name';
 import componentIdToPackageName from '../../utils/bit/component-id-to-package-name';
 import PackageJsonFile from './package-json-file';
+import searchFilesIgnoreExt from '../../utils/fs/search-files-ignore-ext';
+import ComponentVersion from '../../scope/component-version';
 
 /**
  * Add components as dependencies to root package.json
@@ -40,16 +42,14 @@ export async function addComponentsToRoot(consumer: Consumer, components: Compon
 /**
  * Add given components with their versions to root package.json
  */
-export async function addComponentsWithVersionToRoot(consumer: Consumer, componentsIds: BitIds) {
-  const driver = consumer.driver.getDriver(false);
-  const PackageJson = driver.PackageJson;
-
+export async function addComponentsWithVersionToRoot(consumer: Consumer, componentsVersions: ComponentVersion[]) {
   const componentsToAdd = R.fromPairs(
-    componentsIds.map((id) => {
-      return [id.toStringWithoutVersion(), id.version];
+    componentsVersions.map(({ component, version }) => {
+      const packageName = componentIdToPackageName(component.toBitId(), component.bindingPrefix);
+      return [packageName, version];
     })
   );
-  await PackageJson.addComponentsIntoExistingPackageJson(consumer.getPath(), componentsToAdd, npmRegistryName());
+  await _addDependenciesPackagesIntoPackageJson(consumer.getPath(), componentsToAdd);
 }
 
 export async function changeDependenciesToRelativeSyntax(
@@ -156,7 +156,6 @@ export function preparePackageJsonToWrite(
         ...component.testerPackageDependencies
       },
       peerDependencies: component.peerPackageDependencies,
-      componentRootFolder: dir,
       license: `SEE LICENSE IN ${!R.isEmpty(component.license) ? 'LICENSE' : 'UNLICENSED'}`
     });
     packageJsonFile.addDependencies(bitDependencies);
@@ -169,6 +168,8 @@ export function preparePackageJsonToWrite(
     const distRootDir = component.dists.distsRootDir;
     if (!distRootDir) throw new GeneralError('component.dists.distsRootDir is not defined yet');
     distPackageJson = createPackageJsonFile(distRootDir);
+    const distMainFile = searchFilesIgnoreExt(component.dists.get(), component.mainFile, 'relative');
+    distPackageJson.addOrUpdateProperty('main', distMainFile);
   }
 
   return { packageJson, distPackageJson };