+++ src/scope/scope.js
@@ -592,10 +592,7 @@ export default class Scope {
   /**
    * find the components in componentsPool which one of their dependencies include in potentialDependencies
    */
-  async findDirectDependentComponents(
-    componentsPool: BitIds,
-    potentialDependencies: BitIds
-  ): Promise<ComponentVersion[]> {
+  async findDirectDependentComponents(componentsPool: BitIds, potentialDependencies: BitIds): Promise<BitIds> {
     const componentsVersions = await this.loadLocalComponents(componentsPool);
     const dependentsP = componentsVersions.map(async (componentVersion: ComponentVersion) => {
       const component: Version = await componentVersion.getVersion(this.objects);
@@ -605,7 +602,8 @@ export default class Scope {
       return found ? componentVersion : null;
     });
     const dependents = await Promise.all(dependentsP);
-    return removeNils(dependents);
+    const dependentsWithoutNull = removeNils(dependents);
+    return BitIds.fromArray(dependentsWithoutNull.map(c => c.id));
   }
 
   async runComponentSpecs({